Admin Portal/admin-backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
mysql-connector-python==9.0.0
PyJWT==2.9.0
Werkzeug==3.0.3
bcrypt==4.1.2
